The projects I made for my object oriented programming class
That semester we developed a program that is able to process product
orders from an Amazon-like website. In particular, the code will be able
to validate data entered by a user into a form used to place a product
order or review. Web forms are used to collect information from
customers.